@@ -1,6 +1,6 @@
 import 'package:bloc/bloc.dart';
-import 'package:gyawun/ytmusic/ytmusic.dart';
 import 'package:meta/meta.dart';
+import 'package:yt_music/ytmusic.dart';
 
 part 'browse_state.dart';
 
@@ -18,6 +18,28 @@ class BrowseCubit extends Cubit<BrowseState> {
         continuation: feed['continuation'],
         loadingMore: false,
       ));
+    } catch (e, st) {
+      emit(BrowseError(e.toString(), st.toString()));
+    }
+  }
+
+  Future<void> fetchNext() async {
+    final current = state;
+    if (current is! BrowseSuccess) return;
+    if (current.loadingMore || current.continuation == null) return;
+    
+    emit(current.copyWith(loadingMore: true));
+    try {
+      final feed = await _ytMusic.browseContinuation(
+          additionalParams: current.continuation!);
+      emit(
+        BrowseSuccess(
+          header: current.header,
+          sections: [...current.sections, ...feed['sections']],
+          continuation: feed['continuation'],
+          loadingMore: false, 
+        ),
+      );
     } catch (e, st) {
       emit(BrowseError(e.toString(), st.toString()));
     }
