@@ -1,12 +1,18 @@
+import 'dart:ui';
+
+import 'package:fluentui_system_icons/fluentui_system_icons.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_swipe_action_cell/core/cell.dart';
+import 'package:get_it/get_it.dart';
+import 'package:gyawun/core/widgets/expressive_app_bar.dart';
+import 'package:gyawun/core/widgets/song_tile.dart';
+import 'package:gyawun/services/media_player.dart';
+import 'package:gyawun/themes/text_styles.dart';
 
 import '../../../../generated/l10n.dart';
 import '../../../../utils/bottom_modals.dart';
 import '../../../../utils/adaptive_widgets/adaptive_widgets.dart';
-import '../widgets/library_tile.dart';
-import '../widgets/my_playlist_header.dart';
 import 'cubit/favourites_cubit.dart';
 
 class FavouritesPage extends StatelessWidget {
@@ -17,15 +23,12 @@ class FavouritesPage extends StatelessWidget {
     return BlocProvider(
       create: (_) => FavouritesCubit()..load(),
       child: Scaffold(
-        appBar: AppBar(
-          title: Text(S.of(context).Favourites),
-          centerTitle: true,
-        ),
         body: BlocBuilder<FavouritesCubit, FavouritesState>(
           builder: (context, state) {
             return switch (state) {
-              FavouritesLoading() =>
-                const Center(child: AdaptiveProgressRing()),
+              FavouritesLoading() => const Center(
+                child: AdaptiveProgressRing(),
+              ),
               FavouritesError(:final message) => Center(child: Text(message)),
               FavouritesLoaded(:final songs) => _FavouritesBody(songs: songs),
             };
@@ -43,55 +46,136 @@ class _FavouritesBody extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return SingleChildScrollView(
-      child: Center(
-        child: Container(
-          constraints: const BoxConstraints(maxWidth: 1000),
-          padding: const EdgeInsets.symmetric(horizontal: 8),
-          child: Column(
-            children: [
-              MyPlayistHeader(
-                playlist: {'songs': songs},
-              ),
-              Column(
-                children: List.generate(songs.length, (index) {
-                  final song = songs[index];
-
-                  return Padding(
-                    padding: const EdgeInsets.symmetric(vertical: 4),
-                    child: SwipeActionCell(
-                      key: ObjectKey(song),
-                      backgroundColor: Colors.transparent,
-                      trailingActions: [
-                        SwipeAction(
-                          title: S.of(context).Remove,
-                          color: Colors.red,
-                          onTap: (handler) async {
-                            final confirm = await Modals.showConfirmBottomModal(
-                              context,
-                              message: S.of(context).Remove_Message,
-                              isDanger: true,
-                            );
-
-                            if (confirm && context.mounted) {
-                              context
-                                  .read<FavouritesCubit>()
-                                  .remove(song['id'] ?? song['videoId']);
-                            }
-                          },
+    return NestedScrollView(
+      headerSliverBuilder: (context, innerBoxIsScrolled) {
+        return [
+          ExpressiveAppBar(
+            hasLeading: true,
+            child: Column(
+              mainAxisSize: MainAxisSize.min,
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  'Favourites',
+                  maxLines: 1,
+                  overflow: TextOverflow.ellipsis,
+                  style: textStyle(context).copyWith(fontSize: 16),
+                ),
+                SizedBox(height: 2),
+                Text(
+                  S.of(context).nSongs(songs.length),
+                  maxLines: 1,
+                  overflow: TextOverflow.ellipsis,
+                  style: textStyle(
+                    context,
+                  ).copyWith(fontSize: 11, fontWeight: FontWeight.w600),
+                ),
+              ],
+            ),
+          ),
+        ];
+      },
+      body: CustomScrollView(
+        slivers: [
+          SliverToBoxAdapter(
+            child: Padding(
+              padding: const EdgeInsets.all(16.0),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.start,
+                children: [
+                  FilledButton.icon(
+                    style: const ButtonStyle(
+                      padding: WidgetStatePropertyAll(
+                        .symmetric(horizontal: 24, vertical: 16),
+                      ),
+                      shape: WidgetStatePropertyAll(
+                        RoundedRectangleBorder(
+                          borderRadius: .only(
+                            topRight: .circular(8),
+                            bottomRight: .circular(8),
+                            topLeft: .circular(24),
+                            bottomLeft: .circular(24),
+                          ),
                         ),
-                      ],
-                      child: LibraryTile(
-                        songs: songs,
-                        index: index,
                       ),
                     ),
-                  );
-                }),
+                    onPressed: () {
+                      if (songs.isEmpty) return;
+
+                      GetIt.I<MediaPlayer>().playAll(songs);
+                    },
+                    icon: const Icon(FluentIcons.play_24_filled),
+                    label: const Text('Play it'),
+                  ),
+                  SizedBox(width: 4),
+                  FilledButton.tonalIcon(
+                    style: const ButtonStyle(
+                      padding: WidgetStatePropertyAll(
+                        .symmetric(horizontal: 24, vertical: 16),
+                      ),
+                      shape: WidgetStatePropertyAll(
+                        RoundedRectangleBorder(
+                          borderRadius: .only(
+                            topLeft: .circular(8),
+                            bottomLeft: .circular(8),
+                            topRight: .circular(24),
+                            bottomRight: .circular(24),
+                          ),
+                        ),
+                      ),
+                    ),
+
+                    onPressed: () {
+                      if (songs.isEmpty) return;
+                      final shuffled = List.from(songs);
+                      shuffled.shuffle();
+
+                      GetIt.I<MediaPlayer>().playAll(shuffled);
+                    },
+                    icon: const Icon(FluentIcons.arrow_shuffle_24_filled),
+                    label: const Text('Shuffle'),
+                  ),
+                ],
               ),
-            ],
+            ),
           ),
-        ),
+          if (songs.isNotEmpty)
+            SliverList(
+              delegate: SliverChildBuilderDelegate((context, index) {
+                final song = songs[index];
+                return Padding(
+                  padding: const .symmetric(horizontal: 8, vertical: 4),
+                  child: SwipeActionCell(
+                    backgroundColor: Theme.of(
+                      context,
+                    ).colorScheme.surfaceContainerLowest,
+                    key: ObjectKey(song['videoId']),
+                    trailingActions: [
+                      SwipeAction(
+                        title: S.of(context).Remove,
+                        color: Colors.red,
+                        onTap: (handler) async {
+                          final confirm = await Modals.showConfirmBottomModal(
+                            context,
+                            message: S.of(context).Remove_Message,
+                            isDanger: true,
+                          );
+
+                          if (confirm && context.mounted) {
+                            await context.read<FavouritesCubit>().remove(
+                              song['id'] ?? song['videoId'],
+                            );
+                          }
+                        },
+                      ),
+                    ],
+                    child: SongTile(song: song),
+                  ),
+                );
+              }, childCount: songs.length),
+            ),
+          const SliverToBoxAdapter(child: SizedBox(height: 16)),
+        ],
       ),
     );
   }
