@@ -1,14 +1,15 @@
 import 'dart:io';
 
+import 'package:fluentui_system_icons/fluentui_system_icons.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_typeahead/flutter_typeahead.dart';
 import 'package:get_it/get_it.dart';
 import 'package:go_router/go_router.dart';
+import 'package:gyawun/core/widgets/internet_guard.dart';
 import 'package:gyawun/core/utils/service_locator.dart';
 import 'package:gyawun/screens/search/cubit/search_cubit.dart';
-import 'package:gyawun/utils/internet_guard.dart';
 import 'package:loading_indicator_m3e/loading_indicator_m3e.dart';
 
 import '../../../generated/l10n.dart';
@@ -71,14 +72,9 @@ class _SearchPageState extends State<_SearchPage> {
   @override
   Widget build(BuildContext context) {
     return InternetGuard(
-      onInternetLost: () {
-        _focusNode?.unfocus();
-      },
-      onInternetRestored: () {
-        if (widget.title != null) {
-          context.read<SearchCubit>().search('');
-        } else {
-          _focusNode?.requestFocus();
+      onConnectivityRestored:(){
+        if(_textEditingController?.text!=null){
+          context.read<SearchCubit>().search(_textEditingController!.text);
         }
       },
       child: Scaffold(
@@ -108,9 +104,7 @@ class _SearchPageState extends State<_SearchPage> {
                                   maxLines: 1,
                                   autofocus: true,
                                   textInputAction: TextInputAction.search,
-                                  fillColor: Platform.isWindows
-                                      ? null
-                                      : Colors.grey.withValues(alpha: 0.3),
+                                  fillColor: Theme.of(context).colorScheme.surfaceContainer,
                                   contentPadding: const EdgeInsets.symmetric(
                                       vertical: 2, horizontal: 8),
                                   borderRadius: BorderRadius.circular(
@@ -125,7 +119,7 @@ class _SearchPageState extends State<_SearchPage> {
                                         _textEditingController?.text = '';
                                       });
                                     },
-                                    child: const Icon(CupertinoIcons.clear),
+                                    child: const Icon(FluentIcons.dismiss_24_filled),
                                   ),
                                 );
                               },
@@ -174,6 +168,8 @@ class _SearchPageState extends State<_SearchPage> {
         body: BlocBuilder<SearchCubit, SearchState>(
           builder: (context, state) {
             switch (state) {
+              case SearchInitial():
+                return SizedBox.shrink();
               case SearchLoading():
                 return Center(
                   child: LoadingIndicatorM3E(),
