@@ -1,11 +1,14 @@
-import 'dart:io';
-
+import 'package:fluentui_system_icons/fluentui_system_icons.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
+import 'package:gyawun/core/widgets/expressive_app_bar.dart';
+import 'package:gyawun/core/widgets/expressive_list_group.dart';
+import 'package:gyawun/core/widgets/expressive_list_tile.dart';
+import 'package:gyawun/core/widgets/expressive_switch_list_tile.dart';
+import 'package:gyawun/screens/settings/widgets/color_icon.dart';
 
 import '../../../generated/l10n.dart';
-import '../widgets/setting_item.dart';
 import 'cubit/player_settings_cubit.dart';
 
 class PlayerSettingsPage extends StatelessWidget {
@@ -16,54 +19,52 @@ class PlayerSettingsPage extends StatelessWidget {
     return BlocProvider(
       create: (_) => PlayerSettingsCubit(),
       child: Scaffold(
-        appBar: AppBar(
-          title: const Text("Player"),
-        ),
-        body: Center(
-          child: Container(
-            constraints: const BoxConstraints(maxWidth: 1000),
-            child: BlocBuilder<PlayerSettingsCubit, PlayerSettingsState>(
-              builder: (context, state) {
-                final s = state as PlayerSettingsLoaded;
+        body: NestedScrollView(
+          headerSliverBuilder: (context, innerBoxIsScrolled) {
+            return [ExpressiveAppBar(title: "Player", hasLeading: true)];
+          },
+          body: Center(
+            child: Container(
+              constraints: const BoxConstraints(maxWidth: 1000),
+              child: BlocBuilder<PlayerSettingsCubit, PlayerSettingsState>(
+                builder: (context, state) {
+                  final s = state as PlayerSettingsLoaded;
 
-                return ListView(
-                  padding: const EdgeInsets.symmetric(
-                    horizontal: 16,
-                    vertical: 8,
-                  ),
-                  children: [
-                    /// Equalizer
-                    SettingTile(
-                      title: S.of(context).Loudness_And_Equalizer,
-                      leading: const Icon(Icons.equalizer_outlined),
-                      isFirst: true,
-                      isLast: !Platform.isAndroid,
-                      onTap: () {
-                        context.go(
-                          '/settings/player/equalizer',
-                        );
-                      },
+                  return ListView(
+                    padding: const EdgeInsets.symmetric(
+                      horizontal: 16,
+                      vertical: 8,
                     ),
-
-                    /// Skip silence
-                    if (!Platform.isWindows)
-                      SettingSwitchTile(
-                        title: S.of(context).Skip_Silence,
-                        leading: const Icon(
-                          Icons.fast_forward,
-                        ),
-                        value: s.skipSilence,
-                        onChanged: (value) {
-                          context
-                              .read<PlayerSettingsCubit>()
-                              .setSkipSilence(value);
-                        },
-                        isFirst: !Platform.isAndroid,
-                        isLast: true,
+                    children: [
+                      ExpressiveListGroup(
+                        children: [
+                          ExpressiveListTile(
+                            title: Text(S.of(context).Loudness_And_Equalizer),
+                            leading: SettingsColorIcon(
+                              icon: Icons.equalizer_rounded,
+                            ),
+                            trailing: Icon(FluentIcons.chevron_right_24_filled),
+                            onTap: () =>
+                                context.go('/settings/player/equalizer'),
+                          ),
+                          ExpressiveSwitchListTile(
+                            title: Text(S.of(context).Skip_Silence),
+                            leading: SettingsColorIcon(
+                              icon: FluentIcons.fast_forward_24_filled,
+                            ),
+                            value: s.skipSilence,
+                            onChanged: (value) {
+                              context
+                                  .read<PlayerSettingsCubit>()
+                                  .setSkipSilence(value);
+                            },
+                          ),
+                        ],
                       ),
-                  ],
-                );
-              },
+                    ],
+                  );
+                },
+              ),
             ),
           ),
         ),
