@@ -2,9 +2,10 @@ import 'package:bloc/bloc.dart';
 import 'package:flutter/foundation.dart';
 import 'package:get_it/get_it.dart';
 import 'package:hive_flutter/hive_flutter.dart';
+import 'package:yt_music/client.dart';
+import 'package:yt_music/ytmusic.dart';
 
 import '../../../../../../services/settings_manager.dart';
-import '../../../../../../ytmusic/ytmusic.dart';
 
 part 'ytmusic_state.dart';
 
@@ -22,20 +23,22 @@ class YTMusicCubit extends Cubit<YTMusicState> {
   late final VoidCallback _hiveListener;
 
   YTMusicCubit()
-      : super(
-          YTMusicState(
-            location: GetIt.I<SettingsManager>().location,
-            language: GetIt.I<SettingsManager>().language,
-            autofetchSongs: GetIt.I<SettingsManager>().autofetchSongs,
-            streamingQuality: GetIt.I<SettingsManager>().streamingQuality,
-            downloadQuality: GetIt.I<SettingsManager>().downloadQuality,
-            translateLyrics: Hive.box('SETTINGS')
-                .get('TRANSLATE_LYRICS', defaultValue: false),
-            personalisedContent: Hive.box('SETTINGS')
-                .get('PERSONALISED_CONTENT', defaultValue: true),
-            visitorId: Hive.box('SETTINGS').get('VISITOR_ID', defaultValue: ''),
-          ),
-        ) {
+    : super(
+        YTMusicState(
+          location: GetIt.I<SettingsManager>().location,
+          language: GetIt.I<SettingsManager>().language,
+          autofetchSongs: GetIt.I<SettingsManager>().autofetchSongs,
+          streamingQuality: GetIt.I<SettingsManager>().streamingQuality,
+          downloadQuality: GetIt.I<SettingsManager>().downloadQuality,
+          translateLyrics: Hive.box(
+            'SETTINGS',
+          ).get('TRANSLATE_LYRICS', defaultValue: false),
+          personalisedContent: Hive.box(
+            'SETTINGS',
+          ).get('PERSONALISED_CONTENT', defaultValue: true),
+          visitorId: Hive.box('SETTINGS').get('VISITOR_ID', defaultValue: ''),
+        ),
+      ) {
     _settingsListener = _emit;
     _hiveListener = _emit;
 
@@ -54,8 +57,10 @@ class YTMusicCubit extends Cubit<YTMusicState> {
         streamingQuality: _settings.streamingQuality,
         downloadQuality: _settings.downloadQuality,
         translateLyrics: _box.get('TRANSLATE_LYRICS', defaultValue: false),
-        personalisedContent:
-            _box.get('PERSONALISED_CONTENT', defaultValue: true),
+        personalisedContent: _box.get(
+          'PERSONALISED_CONTENT',
+          defaultValue: true,
+        ),
         visitorId: _box.get('VISITOR_ID', defaultValue: ''),
       ),
     );
@@ -73,11 +78,11 @@ class YTMusicCubit extends Cubit<YTMusicState> {
     _settings.autofetchSongs = value;
   }
 
-  void setStreamingQuality(dynamic quality) {
+  void setStreamingQuality(AudioQuality quality) {
     _settings.streamingQuality = quality;
   }
 
-  void setDownloadQuality(dynamic quality) {
+  void setDownloadQuality(AudioQuality quality) {
     _settings.downloadQuality = quality;
   }
 
@@ -87,16 +92,22 @@ class YTMusicCubit extends Cubit<YTMusicState> {
 
   Future<void> setPersonalisedContent(bool value) async {
     await _box.put('PERSONALISED_CONTENT', value);
-    await _ytmusic.resetVisitorId();
+    final config = await YTClient.getConfig();
+    if (config != null) {
+      await _box.put('VISITOR_ID', config.visitorData);
+    }
   }
 
   Future<void> setVisitorId(String id) async {
     await _box.put('VISITOR_ID', id);
-    _ytmusic.refreshHeaders();
+    _ytmusic.updateConfig(visitorData: id);
   }
 
   Future<void> resetVisitorId() async {
-    await _ytmusic.resetVisitorId();
+    final config = await YTClient.getConfig();
+    if (config != null) {
+      await _box.put('VISITOR_ID', config.visitorData);
+    }
   }
 
   @override
